refactor: Move time-update functionility into hook
